requirements.txt
fonttools==4.42.1
freetype-py==2.4.0
pillow==10.0.0
rectpack==0.2.2